requirements.txt
PyQt5==5.15.9
requests==2.31.0
PyYAML==6.0.1
pytest==7.4.3
pytest-qt==4.2.0
